Initial file add
The program started with an initial version that wasn't planned thoroughly and ultimately ended up a jumbled hot mess.  The second file set, ClockActuation was a better planned version using the first attempt as pseudocode.  It was an excellent lesson in a few different things:  Such as always using a UML diagram and pseudocode/flow charts as well as pulling in outside sources cleanly and only once, in this case System time.  Honestly I could have simply started at 12:00 am but wanted the practice parsing and working with <time.h>.