Initial attempt
This is the example of what NOT to do.  I pulled in system time every time I needed a new part instead of pulling it in just once.  I eventually resigned to it being unworkable and started completely over.  This was a valuable lesson in planning UML diagrams and not just "free range" spaghetti programming.  Please see the ClockActuation files.